create_qttube_plugin CMake function, new metadata format
This CMake function greatly simplifies the process of setting up a
proper environment for plugins, especially with the new metadata format.

The new metadata format in question is a dependency on a metadata.json
file in the root project folder of a given plugin, which is processed by
the new CMake function and "loaded" into the program via compile
definitions. It will also be used by the soon-to-come plugin browser.